Started working on tests
